chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
chromadb
pydantic
